Create an apps package
